+++ src/main/kotlin/com/github/magneticflux/rss/namespaces/standard/elements/Enclosure.kt
@@ -1,0 +1,19 @@
+package com.github.magneticflux.rss.namespaces.standard.elements
+
+import com.github.magneticflux.rss.namespaces.standard.converters.EnclosureConverter
+import org.simpleframework.xml.Root
+import java.net.URL
+
+/**
+ * This class represents an enclosure in an [Item].
+ *
+ * @author Mitchell Skaggs
+ * @since 1.0.0
+ * @see EnclosureConverter
+ */
+@Root(name = "enclosure", strict = false)
+data class Enclosure(
+        val url: URL,
+        val length: Long,
+        val type: String
+)
